feat: add license only mode
