+++ infra/compiler/src/index.ts
@@ -3,10 +3,9 @@ import path from 'node:path'
 import url from 'node:url'
 import { parseArgs } from 'node:util'
 
-import chalk from 'chalk'
 import { context } from 'esbuild'
 
-import { checkLicense } from './license.js'
+import { generateLicense } from './license.js'
 import { esbuildProblemMatcherPlugin } from './plugins.js'
 
 import type { PackageJson } from 'type-fest'
@@ -26,6 +25,11 @@ const optionsDef = {
     production: {
         type: 'boolean',
     },
+    onlyLicense: {
+        type: 'boolean',
+        short: 'l',
+        default: false,
+    },
 } as const
 
 const { values: options } = parseArgs({ args, options: optionsDef })
@@ -43,6 +47,18 @@ if (!fss.existsSync(pkgPath)) {
 const pkg = (await import(url.pathToFileURL(pkgPath).href, { with: { type: 'json' } })).default
 
 const absWorkingDir = path.dirname(pkgPath)
+const outdir = path.resolve(absWorkingDir, 'dist')
+
+if (options.onlyLicense) {
+    const metafile = path.join(outdir, 'meta.json')
+    if (!fss.existsSync(metafile)) {
+        throw new Error(`Meta file was not found: ${metafile}\nPlease execute \`pnpm run build\` at root directory.`)
+    }
+    const meta = JSON.parse(fss.readFileSync(metafile, { encoding: 'utf-8' }))
+    generateLicense(rootDir, pkgPath, meta)
+    console.log('The license file was generated successfully.')
+    process.exit(0)
+}
 
 const exports = (pkg.exports || {}) as PackageJson.ExportConditions
 
@@ -61,8 +77,6 @@ for (const [_, exp] of exportedModules) {
 if (entryPoints.length < 1) {
     throw new Error(`No export module found to build at: ${absWorkingDir}`)
 }
-
-const outdir = path.resolve(absWorkingDir, 'dist')
 
 const ctx = await context({
     sourceRoot: absWorkingDir,
@@ -92,20 +106,6 @@ if (options.watch) {
     if (!options.production) {
         fss.writeFileSync(path.join(outdir, 'meta.json'), JSON.stringify(result.metafile))
 
-        const baseLicense = path.join(rootDir, 'LICENSE')
-        const checker = checkLicense(pkgPath, result.metafile)
-        const license = checker.render(baseLicense)
-
-        const existingLicenseText = fss.existsSync(license.file)
-            ? fss.readFileSync(license.file, { encoding: 'utf-8' })
-            : ''
-        if (existingLicenseText !== license.contents) {
-            fss.writeFileSync(license.file, license.contents, { encoding: 'utf-8' })
-            console.info(
-                chalk.yellow(
-                    `\n${path.relative(rootDir, license.file)} was updated. You should commit the updated file.\n`
-                )
-            )
-        }
+        generateLicense(rootDir, pkgPath, result.metafile)
     }
 }
